Show apps on window open.
